chore: remove node-fetch
