+++ package.json
@@ -41,7 +41,6 @@
     "@types/node": "^20.8.7",
     "copyfiles": "^2.4.1",
     "eslint": "^8.52.0",
-    "node-fetch": "^2.7.0",
     "prettier": "^3.0.3",
     "typescript": "^5.2.2"
   }
