+++ src/abi/multicallAbi.ts
@@ -1,0 +1,50 @@
+export const MULTICALL_ABI = [
+  {
+    'inputs': [
+      {
+        'internalType': 'bool',
+        'name': 'requireSuccess',
+        'type': 'bool'
+      },
+      {
+        'components': [
+          {
+            'internalType': 'address',
+            'name': 'target',
+            'type': 'address'
+          },
+          {
+            'internalType': 'bytes',
+            'name': 'callData',
+            'type': 'bytes'
+          }
+        ],
+        'internalType': 'struct Multicall3.Call[]',
+        'name': 'calls',
+        'type': 'tuple[]'
+      }
+    ],
+    'name': 'tryAggregate',
+    'outputs': [
+      {
+        'components': [
+          {
+            'internalType': 'bool',
+            'name': 'success',
+            'type': 'bool'
+          },
+          {
+            'internalType': 'bytes',
+            'name': 'returnData',
+            'type': 'bytes'
+          }
+        ],
+        'internalType': 'struct Multicall3.Result[]',
+        'name': 'returnData',
+        'type': 'tuple[]'
+      }
+    ],
+    'stateMutability': 'payable',
+    'type': 'function'
+  }
+]
