+++ src/abi/omnibridgeAbi.ts
@@ -1,0 +1,7 @@
+export const OMNIBRIDGE_CONTRACT_ABI = [{
+  'inputs': [{'internalType': 'address', 'name': '_nativeToken', 'type': 'address'}],
+  'name': 'bridgedTokenAddress',
+  'outputs': [{'internalType': 'address', 'name': '', 'type': 'address'}],
+  'stateMutability': 'view',
+  'type': 'function'
+}]
