+++ src/scripts/cowFi-tokens.js
@@ -1,4 +1,3 @@
-import fetch from 'node-fetch'
 import fs from 'fs'
 import path from 'path'
 import { backOff } from 'exponential-backoff'
